Generic model moved from package 'mojo' to package 'generic'
