+++ h2o-algos/src/main/java/hex/generic/GenericModelOutput.java
@@ -1,4 +1,4 @@
-package hex.mojo;
+package hex.generic;
 
 import hex.Model;
 import hex.ModelCategory;
