+++ h2o-algos/src/test/java/hex/generic/GenericModelTest.java
@@ -1,4 +1,4 @@
-package hex.mojo;
+package hex.generic;
 
 import hex.ModelCategory;
 import hex.glm.GLM;
